+++ Packages/com.thenathannator.plasticband/CHANGELOG.md
@@ -14,6 +14,8 @@ Various controls did not function correctly due to assumptions I made about how 
 
 - All 5-fret guitars:
   - `strumUp` and `strumDown` aliases on `dpad/up` and `dpad/down` have been removed, and new `strumUp` and `strumDown` controls that replicate `dpad/up` and `dpad/down` have been added in their place.
+- Guitar Hero guitars:
+  - The slider control has been reworked so that instead of being one control with 5 child controls, it's simply one control used 5 times in the main layout. `GuitarHeroSliderSegmentControl` has been removed, `GuitarHeroSliderControl` now inherits from `ButtonControl`, and the `SliderFret` enum previously exposed by it is now private, as it no longer has any public use.
 
 ## [0.1.0] - 2023/15/01